+++ src/app/app.module.ts
@@ -31,7 +31,7 @@ import { EnrolmentPageComponent } from './pages/auth/containers/enrolment-page.c
 import { AuthService } from './pages/auth/services/auth.service'
 import { ClinicalTasksPageComponent } from './pages/clinical-tasks/containers/clinical-tasks-page.component'
 import { FinishPageComponent } from './pages/finish/containers/finish-page.component'
-import { PrepareDataService } from './pages/finish/services/preparedata.service'
+import { PrepareDataService } from './pages/finish/services/prepare-data.service'
 import { TaskCalendarComponent } from './pages/home/components/task-calendar/task-calendar.component'
 import { TaskInfoComponent } from './pages/home/components/task-info/task-info.component'
 import { TaskProgressComponent } from './pages/home/components/task-progress/task-progress.component'
@@ -47,7 +47,7 @@ import { SliderInputComponent } from './pages/questions/components/question/slid
 import { TimedTestComponent } from './pages/questions/components/question/timed-test/timed-test.component'
 import { QuestionsPageComponent } from './pages/questions/containers/questions-page.component'
 import { AnswerService } from './pages/questions/services/answer.service'
-import { AudioRecordService } from './pages/questions/services/audiorecord.service'
+import { AudioRecordService } from './pages/questions/services/audio-record.service'
 import { TimeStampService } from './pages/questions/services/timestamp.service'
 import { ReportPageComponent } from './pages/report/containers/report-page.component'
 import { SettingsPageComponent } from './pages/settings/containers/settings-page.component'
